move metadata functions to better places
