+++ dcpy/builds/load.py
@@ -3,7 +3,8 @@ import typer
 
 from dcpy.utils import postgres, s3
 from dcpy.utils.logging import logger
-from dcpy.connectors.edm import build_metadata, recipes, publishing
+from dcpy.connectors.edm import recipes, publishing
+from dcpy.builds import metadata
 
 
 def setup_build_environments(pg_client: postgres.PostgresClient):
@@ -21,7 +22,7 @@ def load_source_data(recipe_path: Path):
     recipes.write_source_data_versions(recipe_file=Path(recipe_lock_path))
     recipe = recipes.recipe_from_yaml(Path(recipe_lock_path))
 
-    build_name = build_metadata.build_name()
+    build_name = metadata.build_name()
     logger.info(f"Loading source data for {recipe.name} build named {build_name}")
 
     pg_client = postgres.PostgresClient(schema=build_name)
