+++ products/template/build_scripts/__init__.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from dotenv import load_dotenv
 
 from dcpy.utils import postgres
-from dcpy.connectors.edm import build_metadata
+from dcpy.builds import metadata
 
 load_dotenv()
 
@@ -16,7 +16,7 @@ RECIPE_PATH = _product_path / "recipe.yml"
 SQL_QUERY_DIR = _product_path / "sql"
 OUTPUT_DIR = _product_path / "output"
 
-BUILD_NAME = build_metadata.build_name()
+BUILD_NAME = metadata.build_name()
 
 PG_CLIENT = postgres.PostgresClient(
     schema=BUILD_NAME,
